backend/requirements.txt
flask
requests
python-dotenv
inference-sdk
flask_cors